Add a default value parameter to the map method
It currently throws an exception right now when processing a link where the `from` value doesn't exist. This allows the ability to set a default so it's always mapped properly even when the expected value isn't present (some API's return data conditionally).

This also works the other way when reversing (if the `to` value doesn't exist.